+++ engine/src/ElGrapho.js
@@ -50,6 +50,7 @@ ElGrapho.prototype = {
     this.id = UUID.generate();
     this.dirty = true;
     this.hitDirty = true;
+    this.hoverDirty = false;
     this.zoomX = START_SCALE;
     this.zoomY = START_SCALE;
     this.panX = 0;
@@ -93,6 +94,7 @@ ElGrapho.prototype = {
       el.innerHTML = ElGrapho.NumberFormatter.addCommas(index);
     };
     this.hoveredDataIndex = -1;
+    this.selectedIndex = -1;
 
     // all Listeners we need to call remove for on cleanup
     this.allListeners = [];
@@ -107,11 +109,16 @@ ElGrapho.prototype = {
       contextType: 'webgl'
     });
 
+    let hoverLayer = this.hoverLayer = new Concrete.Layer({
+      contextType: '2d'
+    });
+
     let labelsLayer = this.labelsLayer = new Concrete.Layer({
       contextType: '2d'
     });
 
     viewport.add(mainLayer);
+    viewport.add(hoverLayer);
     viewport.add(labelsLayer);
 
 
@@ -284,6 +291,83 @@ ElGrapho.prototype = {
 
     labelsContext.restore();
   },
+  renderRings: function(scale) {
+    let hoverIndex = this.hoveredDataIndex;
+    let selectedIndex = this.selectedIndex;
+
+    if (hoverIndex >= 0 || selectedIndex >= 0) {
+      let halfWidth = this.width/2;
+      let halfHeight = this.height/2;
+
+      
+      // render
+      let scene = this.hoverLayer.scene;
+      let context = scene.context;
+
+
+      context.save();
+      context.translate(this.width/2, this.height/2);
+      context.scale(scale, scale);
+      
+
+      let node, x, y;
+
+      // hover ring
+      if (hoverIndex >= 0) {
+        node = this.model.nodes[hoverIndex];
+        x = (node.x * halfWidth * this.zoomX + this.panX) / scale;
+        y = (node.y * -1 * halfHeight * this.zoomY - this.panY) / scale;
+
+        context.save();
+        if (this.darkMode) {
+          //context.fillStyle = 'rgba(255, 255, 255, 0.4)';   
+          context.strokeStyle = 'white';
+        }
+        else {
+          //context.fillStyle = 'rgba(255, 255, 255, 0.4)';      
+          context.strokeStyle = 'black';
+        }
+
+        context.lineWidth = 3;
+        context.beginPath();
+        context.arc(x, y, 6, 0, 2*Math.PI, false);
+        context.stroke();
+        context.restore();
+      }
+
+      // selected ring
+      if (selectedIndex >= 0) {
+        node = this.model.nodes[selectedIndex];
+        x = (node.x * halfWidth * this.zoomX + this.panX) / scale;
+        y = (node.y * -1 * halfHeight * this.zoomY - this.panY) / scale;
+
+        context.save();
+        if (this.darkMode) {
+          context.strokeStyle = 'white';
+        }
+        else {
+          context.strokeStyle = 'black';      
+        }
+
+        context.lineWidth = 5;
+        context.beginPath();
+        context.arc(x, y, 6, 0, 2*Math.PI, false);
+        context.stroke();
+        context.restore();
+      }
+
+
+
+
+
+      
+      
+
+
+      context.restore();
+    }
+    
+  },
   setDarkMode(darkMode) {
     this.darkMode = darkMode;
 
@@ -454,6 +538,7 @@ ElGrapho.prototype = {
           }
           
           that.hoveredDataIndex = dataIndex;
+          that.hoverDirty = true;
           //that.dirty = true; 
 
           if (that.hoveredDataIndex !== -1) {
@@ -551,9 +636,11 @@ ElGrapho.prototype = {
         let dataIndex = viewport.getIntersection(mousePos.x, mousePos.y);
 
         if (dataIndex === -1) {
+          that.deselectNode();
           that.deselectGroup();
         }
         else {
+          that.selectNode(dataIndex);
           that.selectGroup(that.vertices.points.colors[dataIndex]);
 
           that.fire(Enums.events.NODE_CLICK, {
@@ -582,6 +669,7 @@ ElGrapho.prototype = {
 
         that.dirty = true;
         that.hitDirty = true;
+        that.hoverDirty = true;
       }
     });
 
@@ -737,6 +825,14 @@ ElGrapho.prototype = {
   deselectGroup: function() {
     this.focusedGroup = -1;
     this.dirty = true;
+  },
+  selectNode: function(index) {
+    this.selectedIndex = index;
+    this.hoverDirty = true;
+  },
+  deselectNode: function() {
+    this.selectedIndex = -1;
+    this.hoverDirty = true;
   }
 };